+++ src/Attachment/Template/Generic.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tgallice\FBMessenger\Attachment\Structured;
+namespace Tgallice\FBMessenger\Attachment\Template;
 
-use Tgallice\FBMessenger\Attachment\Structured;
+use Tgallice\FBMessenger\Attachment\Template;
 use Tgallice\FBMessenger\Model\Generic\Element;
 
-class Generic extends Structured
+class Generic extends Template
 {
     const TEMPLATE_TYPE = 'generic';
